update e2fsprogs
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@28913 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/e2fsprogs/lib/uuid/uuid_time.c
@@ -34,6 +34,8 @@
  * %End-Header%
  */
 
+#include "config.h"
+
 #ifdef _WIN32
 #define _WIN32_WINNT 0x0500
 #include <windows.h>
@@ -163,7 +165,7 @@ main(int argc, char **argv)
 		printf("Warning: not a time-based UUID, so UUID time "
 		       "decoding will likely not work!\n");
 	}
-	printf("UUID time is: (%ld, %ld): %s\n", tv.tv_sec, tv.tv_usec,
+	printf("UUID time is: (%ld, %ld): %s\n", tv.tv_sec, (long)tv.tv_usec,
 	       ctime(&time_reg));
 
 	return 0;
